+++ internal/models/airports.go
@@ -23,7 +23,7 @@ func (m *DBModel) GetAirportByID(id string) (Airport, error) {
 	id = strings.Trim(id, " ")
 
 	query := "SELECT icao, iata, name, city, country, elevation, lat, lon " +
-		"FROM airports WHERE icao = ? or iata = ?"
+		"FROM airports_view WHERE icao = ? or iata = ?"
 	row := m.DB.QueryRowContext(ctx, query, id, id)
 
 	err := row.Scan(&airport.ICAO, &airport.IATA, &airport.Name, &airport.City,
@@ -49,7 +49,7 @@ func (m *DBModel) GetAirports() (map[string]Airport, error) {
 	airports := make(map[string]Airport)
 	var airport Airport
 
-	query := "SELECT icao, iata, name, city, country, elevation, lat, lon FROM airports"
+	query := "SELECT icao, iata, name, city, country, elevation, lat, lon FROM airports_view"
 	rows, err := m.DB.QueryContext(ctx, query)
 
 	if err != nil {
@@ -158,4 +158,97 @@ func (m *DBModel) GetAirportCount() (int, error) {
 	}
 
 	return records, nil
+}
+
+// GetStandardAirports returns a list of Standard Airports
+func (m *DBModel) GetStandardAirports() ([]Airport, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var airports []Airport
+	var airport Airport
+
+	query := "SELECT icao, iata, name, city, country, elevation, lat, lon FROM airports"
+	rows, err := m.DB.QueryContext(ctx, query)
+
+	if err != nil {
+		return airports, err
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		err := rows.Scan(&airport.ICAO, &airport.IATA, &airport.Name, &airport.City,
+			&airport.Country, &airport.Elevation, &airport.Lat, &airport.Lon)
+
+		if err != nil {
+			return airports, err
+		}
+
+		airports = append(airports, airport)
+	}
+
+	return airports, nil
+}
+
+// GetCustomAirports returns a list of Custom Airports
+func (m *DBModel) GetCustomAirports() ([]Airport, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var airports []Airport
+	var airport Airport
+
+	query := "SELECT name, city, country, elevation, lat, lon FROM airports_custom"
+	rows, err := m.DB.QueryContext(ctx, query)
+
+	if err != nil {
+		return airports, err
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		err := rows.Scan(&airport.Name, &airport.City,
+			&airport.Country, &airport.Elevation, &airport.Lat, &airport.Lon)
+
+		if err != nil {
+			return airports, err
+		}
+
+		airports = append(airports, airport)
+	}
+
+	return airports, nil
+}
+
+// AddCustomAirport adds new custom/user airport
+func (m *DBModel) AddCustomAirport(arpt Airport) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	query := "INSERT INTO airports_custom " +
+		"(name, city, country, elevation, lat, lon) " +
+		"VALUES (?, ?, ?, ?, ?, ?)"
+	_, err := m.DB.ExecContext(ctx, query,
+		arpt.Name, arpt.City, arpt.Country, arpt.Elevation, arpt.Lat, arpt.Lon,
+	)
+
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// RemoveCustomAirport removes custom/user airport
+func (m *DBModel) RemoveCustomAirport(airport string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := m.DB.ExecContext(ctx, "DELETE FROM airports_custom WHERE name = ?", airport)
+
+	if err != nil {
+		return err
+	}
+
+	return nil
 }